feat: added endpoint for rating songs

+++ api/api/Application/Dtos/RatingDto.cs
@@ -7,5 +7,5 @@ public class RatingDto : IDto
     public long Id { get; set; }
     public required long UserId { get; set; }
     public required long SongId { get; set; }
-    public  required decimal Stars { get; set; }
+    public required int Stars { get; set; }
 }
